+++ src/main/java/com/example/app/WishbujeokApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
