+++ src/main/java/com/example/app/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo.global.config;
+package com.example.app.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
