+++ src/test/java/com/example/app/WishbujeokApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
